+++ RELEASE_NOTES.md
@@ -10,6 +10,8 @@ Changes:
   * Added "Data Approved" and "Data Approver" fields to sequencing runs
   * Added JTT\_7\_GROUPDESC label layout as alternate to JTT\_7 that includes group description
     (detailed sample) instead of the usual extra barcode info
+  * Changed columns shown on Samples, Libraries, and Library Aliquots list pages
+  * Changed warnings on table rows to only show an icon; icon tooltip shows full warning messages
   * Fixed error saving QCs with controls specified
   * Exports initialVolume in Pinery /samples